+++ src/cmd/cmd_c.mli
@@ -28,4 +28,5 @@ val cmd :
   -> entry_point:string option
   -> invoke_with_symbols:bool
   -> out_file:Fpath.t option
+  -> model_out_file:Fpath.t option
   -> unit Result.t