+++ src/cmd/cmd_rust.mli
@@ -25,4 +25,5 @@ val cmd :
   -> invoke_with_symbols:bool
   -> out_file:Fpath.t option
   -> workspace:Fpath.t option
+  -> model_out_file:Fpath.t option
   -> unit Result.t