+++ src/cmd/cmd_zig.ml
@@ -57,8 +57,8 @@ let compile ~workspace ~entry_point ~includes ~out_file debug
 let cmd ~debug ~workers ~includes ~files ~profiling ~unsafe ~optimize
   ~no_stop_at_failure ~no_value ~no_assert_failure_expression_printing
   ~deterministic_result_order ~fail_mode ~concolic ~solver ~profile
-  ~model_format ~entry_point ~invoke_with_symbols ~out_file ~workspace :
-  unit Result.t =
+  ~model_format ~entry_point ~invoke_with_symbols ~out_file ~workspace
+  ~model_out_file : unit Result.t =
   let* workspace =
     match workspace with
     | Some path -> Ok path
@@ -83,4 +83,4 @@ let cmd ~debug ~workers ~includes ~files ~profiling ~unsafe ~optimize
     ~profiling ~debug ~unsafe ~rac:false ~srac:false ~optimize ~workers
     ~no_stop_at_failure ~no_value ~no_assert_failure_expression_printing
     ~deterministic_result_order ~fail_mode ~workspace ~solver ~files ~profile
-    ~model_format ~entry_point ~invoke_with_symbols
+    ~model_format ~entry_point ~invoke_with_symbols ~model_out_file