+++ CHANGES.md
@@ -33,6 +33,7 @@
 - add `--invoke-with-symbols` option
 - add `owi iso` subcommand
 - add `--workspace` option
+- add `--model-out-file` option
 
 ## 0.2 - 2024-04-24